changed way of working with color service

+++ src/Services/Kros.Tags.Api/Application/Commands/UpdateTag/UpdateTagCommandHandler.cs
@@ -40,7 +40,7 @@ namespace Kros.Tags.Api.Application.Commands
         /// <inheritdoc />
         public async Task<Unit> Handle(UpdateTagCommand request, CancellationToken cancellationToken)
         {
-            var colors = _colorManagementService.GetUsedColors(request.OrganizationId);
+            /*var colors = _colorManagementService.GetUsedColors(request.OrganizationId);
             if (request.ColorARGBValue == 0)
             {
                 request.ColorARGBValue = RandomColor.GetColor(ColorScheme.Random, Luminosity.Bright).ToArgb();
@@ -61,6 +61,19 @@ namespace Kros.Tags.Api.Application.Commands
                 var colorValue = RandomColor.GetColor(ColorScheme.Random, Luminosity.Bright).ToArgb();
                 request.ColorARGBValue = colorValue;
                 colorExistsInStorage = colors.Any(c => c.ColorValue == request.ColorARGBValue.ToString());
+            }*/
+
+            var generatedColor = _colorManagementService.CheckAndGenerateColor(
+                request.OrganizationId, 
+                request.ColorARGBValue, 
+                request.OldColorARGBValue);
+            if (generatedColor == 0)
+            {
+                throw new InvalidOperationException("Color already exist in storage.");
+            }
+            else
+            {
+                request.ColorARGBValue = generatedColor;
             }
 
             await _colorManagementService.DeleteColor(request.OldColorARGBValue, request.OrganizationId);
